pyproject.toml
[project]
name = "super-vault"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "prefect>=3.2.1",
    "clickhouse-driver>=0.2.1",
    "pandas>=2.1.1",
    "python-dotenv>=1.0.1",
    "prefect-dbt>=0.1.1",
    "requests>=2.0.0",
    "web3>=7.8.0"
]
